Homework 2 - VBA
This may be incomplete. The TAs are aware I had a bit of a bizarre week this week and had to play a lot of catch up. I'd like to move forward and focus on python and staying on track/focusing on getting that homework started this weekend. I wanted to show that I did try. I listened to Kevin's lecture regarding the homework and that helped a significant amount. I need to do some more work with VBA on my own time and plan to because I do a lot of excel work at my job (but python will probably help me do more/better things with the data at my job)